api-agentic-sistem-test1/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-community
chromadb
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pypdf
tiktoken
